pyproject.toml
[tool.poetry]
name = "todo"
version = "0.1.0"
description = "A simple todo web app."
authors = ["Nurbakyt Madibek <madibekov.nurbakhyt@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.7"
pre-commit = "^2.11.1"
ipython = "^7.21.0"
djangorestframework = "^3.12.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
